requirements.txt
Flask==2.2.3
torch==2.0.0
torchvision==0.15.1
Pillow==9.5.0
